Refactor gamefi page components
