+++ src/views/Gamefi/NewUI/InputComponent.tsx
