+++ src/views/Gamefi/NewUI/types.ts
@@ -1,0 +1,5 @@
+export type Series = {
+    name: string
+    data: number[]
+  }
+
